Revert "Add snapshot limiting"
