+++ dropbox-upload/config.json
@@ -13,15 +13,11 @@
     ],
     "options": {
         "access_token": "<YOUR_ACCESS_TOKEN>",
-        "dropbox_dir": "<DROPBOX_UPLOAD_LOCATION>",
-        "keep": 10,
-        "debug": false
+        "dropbox_dir": "<DROPBOX_UPLOAD_LOCATION>"
     },
     "schema": {
         "access_token": "str",
         "dropbox_dir": "str",
-        "keep": "int?",
-        "mins_between_backups": "int?",
         "debug": "bool?"
     }
 }
