--- dropbox-upload/dropbox_upload/util.py
@@ -1,8 +1,0 @@
-def bytes_to_human(nbytes):
-    suffixes = ["B", "KB", "MB", "GB", "TB", "PB"]
-    i = 0
-    while nbytes >= 1024 and i < len(suffixes) - 1:
-        nbytes /= 1024.
-        i += 1
-    f = ("%.2f" % nbytes).rstrip("0").rstrip(".")
-    return "%s %s" % (f, suffixes[i])
