+++ tests/test_dropbox_upload.py
@@ -2,55 +2,56 @@ import json
 import logging
 import pathlib
 
-from dropbox_upload import backup, config, dropbox, hassio, limit, util
+import upload
+from dropbox import exceptions
 
 
 def test_bytes_to_human():
-    assert util.bytes_to_human(1024) == "1 KB"
+    assert upload.bytes_to_human(1024) == "1 KB"
 
 
 def test_compute_dropbox_hash(tmpdir):
     p = tmpdir.join("fake-snapshot.txt")
     p.write("fake content")
-    assert dropbox.compute_dropbox_hash(p.strpath).startswith("de92bfc3d2")
+    assert upload.compute_dropbox_hash(p.strpath).startswith("de92bfc3d2")
 
 
 def test_load_config(tmpdir):
     p = tmpdir.join("config.json")
     p.write(json.dumps({}))
-    assert config.load_config(p.strpath) == {}
+    assert upload.load_config(p.strpath) == {}
 
 
 def test_local_path():
     expected = pathlib.Path("/backup/SLUG.tar")
-    assert backup.local_path({"slug": "SLUG"}) == expected
+    assert upload.local_path({"slug": "SLUG"}) == expected
 
 
 def test_dropbox_path():
     dropbox_dir = pathlib.Path("/dropbox_dir/")
     expected = pathlib.Path("/dropbox_dir/SLUG.tar")
-    assert backup.dropbox_path(dropbox_dir, {"slug": "SLUG"}) == expected
+    assert upload.dropbox_path(dropbox_dir, {"slug": "SLUG"}) == expected
 
 
 def test_setup_logging_normal():
-    logger = config.setup_logging({})
+    logger = upload.setup_logging({})
     assert logger.level == logging.INFO
 
 
 def test_setup_logging_debug():
-    logger = config.setup_logging({"debug": True})
+    logger = upload.setup_logging({"debug": True})
     assert logger.level == logging.DEBUG
 
 
 def test_hassio_api(requests_mock):
     requests_mock.get("http://hassio/snapshots", text=json.dumps({"data": {}}))
-    assert hassio.hassio_get("snapshots") == {}
+    assert upload.hassio_get("snapshots") == {}
 
 
 def test_list_snapshots(requests_mock):
     data = {"data": {"snapshots": []}}
     requests_mock.get("http://hassio/snapshots", text=json.dumps(data))
-    assert hassio.list_snapshots() == []
+    assert upload.list_snapshots() == []
 
 
 def test_backup_no_snapshots(tmpdir, requests_mock, caplog):
@@ -58,68 +59,60 @@ def test_backup_no_snapshots(tmpdir, requests_mock, caplog):
     # Create config file
     p = tmpdir.join("config.json")
     p.write(json.dumps({"dropbox_dir": "snapshots", "access_token": "token"}))
-    cfg = config.load_config(p.strpath)
+    config = upload.load_config(p.strpath)
 
-    backup.backup(None, cfg, [])
+    upload.backup(None, config, [])
 
     assert (
-        "dropbox_upload.backup",
+        "docker_upload",
         logging.WARNING,
         "No snapshots found to backup",
     ) in caplog.record_tuples
 
 
-def test_limit_snapshots_no_setting(cfg, caplog):
-    cfg["keep"] = None
-    limit.limit_snapshots(None, cfg, [])
+def test_main_invalid_token(tmpdir, requests_mock, caplog):
+
+    # create config file
+    p = tmpdir.join("config.json")
+    p.write(json.dumps({"dropbox_dir": "snapshots", "access_token": "token"}))
+
+    requests_mock.post(
+        "https://api.dropboxapi.com/2/users/get_current_account",
+        exc=exceptions.AuthError("Request ID", "Invalid token"),
+    )
+
+    upload.main(p.strpath)
+
     assert (
-        "dropbox_upload.limit",
-        logging.WARNING,
-        "keep not set. We wont remove old snapshots",
+        "docker_upload",
+        logging.ERROR,
+        "Invalid access token",
     ) in caplog.record_tuples
 
 
-def test_limit_snapshots_zero(cfg, caplog):
-    cfg["keep"] = 0
-    limit.limit_snapshots(None, cfg, [])
-    assert (
-        "dropbox_upload.limit",
-        logging.WARNING,
-        "keep not set. We wont remove old snapshots",
-    ) in caplog.record_tuples
+def test_main(tmpdir, requests_mock, caplog):
 
+    # create config file
+    p = tmpdir.join("config.json")
+    p.write(json.dumps({"dropbox_dir": "snapshots", "access_token": "token"}))
 
-def test_limit_snapshots_not_reached(cfg, caplog):
-    limit.limit_snapshots(None, cfg, [])
+    # Mock hass API
+    data = {"data": {"snapshots": []}}
+    requests_mock.get("http://hassio/snapshots", text=json.dumps(data))
+
+    class DropboxAPI:
+        def __init__(self, *args, **kwargs):
+            pass
+
+        def users_get_current_account(self):
+            return
+
+    upload.main(p.strpath, sleeper=lambda x: True, DropboxAPI=DropboxAPI)
+
     assert (
-        "dropbox_upload.limit",
-        logging.WARNING,
-        "keep not set. We wont remove old snapshots",
-    ) not in caplog.record_tuples
-    assert (
-        "dropbox_upload.limit",
+        "docker_upload",
         logging.INFO,
-        "Not reached the maximum number of snapshots",
+        "Starting Snapshot backup",
     ) in caplog.record_tuples
 
-
-def test_limit_snapshots(cfg, caplog, snapshots, requests_mock, dropbox_fake):
-    requests_mock.post("http://hassio/snapshots/slug2/remove", text='{"data":null}')
-    requests_mock.post("http://hassio/snapshots/slug3/remove", text='{"data":null}')
-    cfg["keep"] = 2
-    limit.limit_snapshots(dropbox_fake(), cfg, snapshots)
-    assert (
-        "dropbox_upload.limit",
-        logging.WARNING,
-        "keep not set. We wont remove old snapshots",
-    ) not in caplog.record_tuples
-    assert (
-        "dropbox_upload.limit",
-        logging.INFO,
-        "Not reached the maximum number of snapshots",
-    ) not in caplog.record_tuples
-    assert (
-        "dropbox_upload.limit",
-        logging.INFO,
-        "Deleting 2 snapshots",
-    ) in caplog.record_tuples
+    assert ("docker_upload", logging.INFO, "Uploads complete") in caplog.record_tuples
