+++ tox.ini
@@ -9,8 +9,8 @@ deps =
   -rdropbox-upload/requirements.txt
   -rtest-requirements.txt
 commands=
-  pytest -v --basetemp={envtmpdir} \
-    --cov=dropbox_upload --cov-fail-under=65 --cov-report=term-missing {posargs}
+  pytest --basetemp={envtmpdir} \
+    --cov=upload --cov-fail-under=60 --cov-report=term-missing {posargs}
 
 [testenv:format]
 basepython=python3
